take SigHash instead of raw byte in hashForTaprootSignature

+++ core/src/test/java/org/bitcoinj/core/TransactionTest.java
@@ -549,7 +549,7 @@ public class TransactionTest {
                 inputIndex,
                 tapLeafHash,
                 parseTaprootPrevOuts(),
-                sigHashType.byteValue()
+                sigHashType
         );
 
         assertArrayEquals(sigHash.getBytes(), Hex.decode(expectedSigHashHex));
